+++ package.json
@@ -27,7 +27,6 @@
     "express-mysql-session": "^3.0.0",
     "express-session": "^1.17.3",
     "express-useragent": "^1.0.15",
-    "lodash": "^4.17.21",
     "method-override": "^3.0.0",
     "multer": "^1.4.5-lts.1",
     "mybatis-mapper": "^0.6.8",
