+++ src/pages/changePassword/index.ts
@@ -1,0 +1,3 @@
+import ChangePassword from './ChangePassword';
+
+export default ChangePassword;
