+++ src/pages/editThread/index.ts
@@ -1,0 +1,3 @@
+import EditThread from './EditThread';
+
+export default EditThread;
