+++ src/pages/join/index.ts
@@ -1,0 +1,3 @@
+import Join from './Join';
+
+export default Join;
