+++ src/pages/notfound/index.ts
@@ -1,0 +1,3 @@
+import NotFound from './NotFound';
+
+export default NotFound;
