+++ src/pages/thread/index.ts
@@ -1,0 +1,3 @@
+import Thread from './Thread';
+
+export default Thread;
