+++ src/pages/userprofile/UserProfile.tsx
@@ -16,15 +16,15 @@ import { Link, useParams } from 'react-router-dom';
 import styled from 'styled-components';
 import Swal from 'sweetalert2';
 
-import { deleteUser, editDescription, getUserProfile, uploadUserProfile } from '../api/userApi';
-import BackPageIcon from '../components/BackPageIcon';
-import Header from '../components/Header';
-import Spinner from '../components/Spinner';
-import useError from '../hooks/useError';
-import useLoginInfo from '../hooks/useLoginInfo';
-import Meta from '../Meta';
-import { media } from '../styles/mediaQuery';
-import { IActivityCount, Iprofile } from '../types/user';
+import { deleteUser, editDescription, getUserProfile, uploadUserProfile } from '../../api/userApi';
+import BackPageIcon from '../../components/BackPageIcon';
+import Header from '../../components/Header';
+import Spinner from '../../components/Spinner';
+import useError from '../../hooks/useError';
+import useLoginInfo from '../../hooks/useLoginInfo';
+import Meta from '../../Meta';
+import { media } from '../../styles/mediaQuery';
+import { IActivityCount, Iprofile } from '../../types/user';
 
 const Grid = styled.section`
   margin: auto;
