+++ src/pages/userprofile/index.ts
@@ -1,0 +1,3 @@
+import UserProfile from './UserProfile';
+
+export default UserProfile;
