+++ src/pages/write/index.ts
@@ -1,0 +1,3 @@
+import Write from './Write';
+
+export default Write;
